// src/teacher/interfaces/teacher.interface.ts
export interface TeacherInterface {
  teacherId?: string
  firstName?: string
  lastName?: string
  contactNumber?: string
  email?: string
  gender?: string
  socialCategory?: string
  birthDate?: string
  designation?: string
  cadre?: string
  profQualification?: string
  joiningDate?: string
  subjectId?: string
  bloodGroup?: string
  maritalStatus?:string
  blockI?: string
  address?: string
  compSkills?: string
  disability?: string
  religion?: string
  homeDistance?: string
  roles?: string
  schoolId?: string
  acrId?: string
  retirementDate?: string
  workingStatus?: string
  osCreatedAt?: string
  osUpdatedAt?: string
  osCreatedBy?: string
  osUpdatedBy?: string
}